pyproject.toml
[tool.poetry]
name = "aussieshopper"
version = "1.0.0"
description = ""
authors = ["Alaa Salman <alaa@codedemigod.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "~=2.2"
feedparser = "=~6.0"
python-telegram-bot = "~=6.0"
gunicorn = "~=20.0"
djangorestframework = "~=3.6"
celery = "~=4.0"
redis = "~=3.4"
pygments = "==2.2.0"
django-extensions = "~=3.1"
ipython = "~=7.19"
psycopg2-binary = "~=2.8"
django-webpack-loader = "~=0.7"
sentry-sdk = "=~1.5"
django-anymail = {version = "7.2.1", extras = ["sendgrid"]}
django-simple-captcha = "^0.5.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
